Fix blueprint endpoint references in templates
