+++ README.md
@@ -14,6 +14,27 @@ export DATABASE_URL="postgres://user:password@hostname:5432/dbname"
 
 If `DATABASE_URL` is not provided, a local `app.db` SQLite file will be used.
 
+## Environment Variables
+
+Several settings are loaded from environment variables so you do not need to
+store secrets in the source code:
+
+* `SECRET_KEY` &ndash; Flask session secret.
+* `API_KEY` &ndash; Financial Modeling Prep API key (required).
+* `SMTP_SERVER`, `SMTP_PORT`, `SMTP_USERNAME`, `SMTP_PASSWORD` &ndash; SMTP
+  credentials for sending alert emails.
+
+Example:
+
+```bash
+export SECRET_KEY="supersecret"
+export API_KEY="your_fmp_api_key"
+export SMTP_SERVER="smtp.example.com"
+export SMTP_PORT=587
+export SMTP_USERNAME="user@example.com"
+export SMTP_PASSWORD="password"
+```
+
 ## Setup
 
 Install dependencies:
@@ -27,6 +48,10 @@ Run the application:
 ```bash
 python app.py
 ```
+
+The Flask code is now organized as a package under `app/` using Blueprints.
+The `app.py` file simply creates the application via `create_app()` and starts
+the server.
 
 ## Custom P/E Thresholds
