+++ templates/index.html
@@ -13,14 +13,14 @@
         <div class="container">
             {% if current_user.is_authenticated %}
                 <span class="navbar-text me-2">Logged in as {{ current_user.username }}</span>
-                <a href="{{ url_for('watchlist') }}" class="btn btn-sm btn-outline-primary me-2">Watchlist</a>
-                <a href="{{ url_for('favorites') }}" class="btn btn-sm btn-outline-primary me-2">Favorites</a>
-                <a href="{{ url_for('export_history') }}" class="btn btn-sm btn-outline-secondary me-2">Export History</a>
-                <a href="{{ url_for('settings') }}" class="btn btn-sm btn-outline-secondary me-2">Settings</a>
-                <a href="{{ url_for('logout') }}" class="btn btn-sm btn-danger">Logout</a>
+                <a href="{{ url_for('user.watchlist') }}" class="btn btn-sm btn-outline-primary me-2">Watchlist</a>
+                <a href="{{ url_for('user.favorites') }}" class="btn btn-sm btn-outline-primary me-2">Favorites</a>
+                <a href="{{ url_for('user.export_history') }}" class="btn btn-sm btn-outline-secondary me-2">Export History</a>
+                <a href="{{ url_for('user.settings') }}" class="btn btn-sm btn-outline-secondary me-2">Settings</a>
+                <a href="{{ url_for('auth.logout') }}" class="btn btn-sm btn-danger">Logout</a>
             {% else %}
-                <a href="{{ url_for('login') }}" class="btn btn-sm btn-outline-primary me-2">Login</a>
-                <a href="{{ url_for('signup') }}" class="btn btn-sm btn-outline-secondary">Sign Up</a>
+                <a href="{{ url_for('auth.login') }}" class="btn btn-sm btn-outline-primary me-2">Login</a>
+                <a href="{{ url_for('auth.signup') }}" class="btn btn-sm btn-outline-secondary">Sign Up</a>
             {% endif %}
         </div>
     </nav>
@@ -48,7 +48,7 @@
                                         <li class="list-group-item">{{ item }}</li>
                                     {% endfor %}
                                     </ul>
-                                    <a href="{{ url_for('clear_history') }}" class="btn btn-sm btn-danger mt-2">Clear History</a>
+                                    <a href="{{ url_for('main.clear_history') }}" class="btn btn-sm btn-danger mt-2">Clear History</a>
                                 </div>
                             {% endif %}
 
@@ -112,8 +112,8 @@
                                 <p><strong>Earnings Growth:</strong> {{ earnings_growth }}%</p>
                             {% endif %}
                             {% if current_user.is_authenticated and symbol %}
-                                <a href="{{ url_for('add_watchlist', symbol=symbol) }}" class="btn btn-sm btn-outline-primary mb-2">Add to Watchlist</a>
-                                <a href="{{ url_for('add_favorite', symbol=symbol) }}" class="btn btn-sm btn-outline-primary mb-2">Add to Favorites</a>
+                                <a href="{{ url_for('user.add_watchlist', symbol=symbol) }}" class="btn btn-sm btn-outline-primary mb-2">Add to Watchlist</a>
+                                <a href="{{ url_for('user.add_favorite', symbol=symbol) }}" class="btn btn-sm btn-outline-primary mb-2">Add to Favorites</a>
                             {% endif %}
 
                             {% if history_prices %}
@@ -152,8 +152,8 @@
                                 </script>
                             {% endif %}
                             <div class="mt-3">
-                                <a href="{{ url_for('download', symbol=symbol, format='csv') }}" class="btn btn-secondary me-2">Download CSV</a>
-                                <a href="{{ url_for('download', symbol=symbol, format='pdf') }}" class="btn btn-secondary">Download PDF</a>
+                                <a href="{{ url_for('main.download', symbol=symbol, format='csv') }}" class="btn btn-secondary me-2">Download CSV</a>
+                                <a href="{{ url_for('main.download', symbol=symbol, format='pdf') }}" class="btn btn-secondary">Download PDF</a>
                             </div>
                         {% endif %}
                     </div>
