+++ templates/login.html
@@ -31,7 +31,7 @@
                             </div>
                         </form>
                         <div class="mt-3 text-center">
-                            <a href="{{ url_for('signup') }}">Sign Up</a>
+                            <a href="{{ url_for('auth.signup') }}">Sign Up</a>
                         </div>
                     </div>
                 </div>
