+++ templates/signup.html
@@ -35,7 +35,7 @@
                             </div>
                         </form>
                         <div class="mt-3 text-center">
-                            <a href="{{ url_for('login') }}">Login</a>
+                            <a href="{{ url_for('auth.login') }}">Login</a>
                         </div>
                     </div>
                 </div>
