+++ templates/watchlist.html
@@ -29,7 +29,7 @@
                     <div class="input-group" style="width: 200px;">
                         <input type="number" step="0.01" name="threshold" class="form-control form-control-sm" value="{{ item.pe_threshold }}">
                         <button class="btn btn-sm btn-primary" type="submit">Update</button>
-                        <a href="{{ url_for('delete_watchlist_item', item_id=item.id) }}" class="btn btn-sm btn-danger ms-1">Delete</a>
+                        <a href="{{ url_for('user.delete_watchlist_item', item_id=item.id) }}" class="btn btn-sm btn-danger ms-1">Delete</a>
                     </div>
                 </form>
             </li>
@@ -38,8 +38,8 @@
         {% else %}
         <p>No watchlist items.</p>
         {% endif %}
-    <a href="{{ url_for('index') }}" class="btn btn-secondary mt-3 me-2">Back</a>
-    <a href="{{ url_for('favorites') }}" class="btn btn-secondary mt-3">Favorites</a>
+    <a href="{{ url_for('main.index') }}" class="btn btn-secondary mt-3 me-2">Back</a>
+    <a href="{{ url_for('user.favorites') }}" class="btn btn-secondary mt-3">Favorites</a>
     </div>
     <script>
         if ('serviceWorker' in navigator) {
